+++ packages/docs/src/components/BorderRadiosList.tsx
@@ -18,15 +18,15 @@ export const BorderRadiosList: FC = () => {
   ]
 
   return (
-    <div className="flex flex-wrap p-4 bg-sb-bg-light gap-4 rounded-md">
+    <div className="flex flex-wrap gap-4 rounded-md bg-sb-bg-light p-4">
       {pickedOptions.map(({ cls, property, comment, custom }) => (
-        <div key={cls} className="flex flex-col sb-unstyled">
-          <div className={cn(['w-24 h-24 bg-primary-500', cls])} />
-          <span className="text-gray-900 font-bold text-sm mt-2">
+        <div key={cls} className="sb-unstyled flex flex-col">
+          <div className={cn(['h-24 w-24 bg-primary-500', cls])} />
+          <span className="mt-2 text-sm font-bold text-gray-900">
             Class: {custom && '*'}
             {cls}
           </span>
-          <small className="text-gray-600 text-xs font-semibold">
+          <small className="text-xs font-semibold text-gray-600">
             Property: {property} &nbsp;
             <br />
             {comment && <em className=" text-[10px]">{` /* ${comment} */`}</em>}