+++ packages/docs/src/components/FontFamiliesList.tsx
@@ -2,101 +2,92 @@ import { FC } from 'react'
 import { cn } from '@click-tools-ui/react'
 
 const styles = {
-  list: cn('flex bg-sb-bg-light gap-6 rounded-md flex-wrap'),
-  box: cn('p-6  flex flex-col'),
+  list: cn('flex flex-wrap gap-6 rounded-md bg-sb-bg-light'),
+  box: cn('flex flex-col p-6'),
   'box-header': cn(
-    'sb-unstyled mb-12 flex justify-between items-center flex-wrap-reverse gap-3',
+    'sb-unstyled mb-12 flex flex-wrap-reverse items-center justify-between gap-3',
   ),
-  headingBase: cn('sb-unstyled text-gray-600 text-2xl leading-snug font-light'),
-  textBase: cn('text-3xl font-medium text-gray-700 sb-unstyled'),
+  headingBase: cn(
+    'sb-unstyled font-roboto text-2xl font-light leading-snug text-gray-600',
+  ),
+  textBase: cn('sb-unstyled font-roboto text-3.5xl font-medium text-gray-700'),
   fontText: (className: string | string[]) =>
-    cn('text-gray-700 text-7xl !leading-normal sb-unstyled', className),
-  body: cn('flex flex-col gap-3 mb-14 sb-unstyled'),
-  cardDoc: cn('p-3 bg-gray-200 rounded-sm w-fit sb-unstyled'),
+    cn('sb-unstyled text-7xl !leading-normal text-gray-700', className),
+  body: cn('sb-unstyled mb-14 flex flex-col gap-3'),
+  cardDoc: cn('sb-unstyled w-fit rounded-sm bg-gray-200 p-3'),
   footer: cn('flex flex-col gap-10'),
 }
 
+type FontFamilyDescription = {
+  heading: string
+  family: string
+  familyClass: string
+  sourceURL: string
+  fontWeights: string
+  lineHeight: string
+  fontSizes: string
+}
+
 export const FontFamiliesList: FC = () => {
+  const fontsFamilies: FontFamilyDescription[] = [
+    {
+      heading: 'Heading / Title',
+      family: 'Mont Serrat',
+      familyClass: 'font-montserrat',
+      sourceURL: 'https://fonts.google.com/specimen/Montserrat',
+      fontWeights: 'Bold & Black',
+      lineHeight: '130%',
+      fontSizes: '24, 32, 48, 72px',
+    },
+    {
+      heading: 'Text / Paragraph',
+      family: 'Roboto',
+      familyClass: 'font-roboto',
+      sourceURL: 'https://fonts.google.com/?query=roboto',
+      fontWeights: 'Regular & Medium & Bold',
+      lineHeight: '130% & 160%',
+      fontSizes: '12,16, 24 px',
+    },
+  ]
+
   return (
     <div className={styles.list}>
-      {/* Roboto */}
-      <div className={styles.box}>
-        <div className={styles.body}>
-          <h3 className={styles.headingBase}>Font Family - Text</h3>
-          <span className={styles.fontText('font-roboto font-bold')}>
-            Roboto
-          </span>
+      {fontsFamilies.map((f) => (
+        <div key={f.family} className={styles.box}>
+          <div className={styles.body}>
+            <h3 className={styles.headingBase}>Font Family - {f.heading}</h3>
+            <span className={styles.fontText(['font-bold', f.familyClass])}>
+              {f.family}
+            </span>
 
-          <div className={styles.cardDoc}>
-            <span className="font-medium">The font can be accessed by </span>
-            <strong className="underline">
-              <a
-                href="https://fonts.google.com/?query=roboto"
-                target="_blank"
-                rel="noreferrer"
-              >
-                Google Fonts
-              </a>
-            </strong>
+            <div className={styles.cardDoc}>
+              <span className="font-medium">The font can be accessed by </span>
+              <strong className="underline">
+                <a href={f.sourceURL} target="_blank" rel="noreferrer">
+                  Google Fonts
+                </a>
+              </strong>
+            </div>
+          </div>
+
+          <div className={styles.footer}>
+            <div>
+              <h3 className={styles.headingBase}>Font Weights</h3>
+              <span className={styles.textBase}>{f.fontWeights}</span>
+            </div>
+
+            <div>
+              <h3 className={styles.headingBase}>Line Height</h3>
+              <span className={styles.textBase}>{f.lineHeight}</span>
+            </div>
+
+            <div>
+              <h3 className={styles.headingBase}>Font Sizes</h3>
+              <span className={styles.textBase}>{f.fontSizes}</span>
+            </div>
           </div>
         </div>
-
-        <div className={styles.footer}>
-          <div>
-            <h3 className={styles.headingBase}>Font Weights</h3>
-            <span className={styles.textBase}>Regular & Medium & Bold</span>
-          </div>
-
-          <div>
-            <h3 className={styles.headingBase}>Line Height</h3>
-            <span className={styles.textBase}>130%</span>
-          </div>
-
-          <div>
-            <h3 className={styles.headingBase}>Font Sizes</h3>
-            <span className={styles.textBase}>12,14,16,20 & 24 px</span>
-          </div>
-        </div>
-      </div>
-      {/* MontSerrat */}
-      <div className={styles.box}>
-        <div className={styles.body}>
-          <h3 className={styles.headingBase}>Font Family - Heading</h3>
-          <span className={styles.fontText('font-montserrat  font-bold')}>
-            Mont Serrat
-          </span>
-
-          <div className={styles.cardDoc}>
-            <span className="font-medium">The font can be accessed by </span>
-            <strong className="underline">
-              <a
-                href="https://fonts.google.com/specimen/Montserrat?query=montserrat"
-                target="_blank"
-                rel="noreferrer"
-              >
-                Google Fonts
-              </a>
-            </strong>
-          </div>
-        </div>
-
-        <div className={styles.footer}>
-          <div>
-            <h3 className={styles.headingBase}>Font Weights</h3>
-            <span className={styles.textBase}>Semibold & Bold</span>
-          </div>
-
-          <div>
-            <h3 className={styles.headingBase}>Line Height</h3>
-            <span className={styles.textBase}>100% & 130%</span>
-          </div>
-
-          <div>
-            <h3 className={styles.headingBase}>Font Sizes</h3>
-            <span className={styles.textBase}>12,14,16,20 & 24 px</span>
-          </div>
-        </div>
-      </div>
+      ))}
     </div>
   )
 }