+++ packages/react/src/components/Button/Button.spec.tsx
@@ -11,12 +11,9 @@ describe('Button Component', () => {
     expect(button).toHaveTextContent('Hello')
     expect(button.type).toBe('button')
     expect(button).toBeEnabled()
-    expect(button).toHaveClass(
-      btnStyles({ variant: 'primary', size: 'md', rounded: 'full' }),
-      {
-        exact: true,
-      },
-    )
+    expect(button).toHaveClass(btnStyles({ variant: 'primary', size: 'md' }), {
+      exact: true,
+    })
   })
 
   it('should be able to add new classNames', () => {
@@ -25,7 +22,7 @@ describe('Button Component', () => {
     const button = screen.getByRole('button') as HTMLButtonElement
 
     expect(button).toHaveClass(
-      btnStyles({ variant: 'primary', size: 'md', rounded: 'full', className }),
+      btnStyles({ variant: 'primary', size: 'md', className }),
       {
         exact: true,
       },