+++ packages/react/src/index.tsx
@@ -6,3 +6,5 @@ export * from '~/components/form/Input'
 export * from '~/components/typography'
 
 export * from '~/utils/cn'
+
+export * from '~/components/typography/index'